optimizando comando profile

+++ src/data/commands/profile.js
@@ -1,6 +1,24 @@
 import { MessageEmbed } from 'discord.js';
 
-import { sendMsg } from '../util';
+import { sendMsg, getUserWithId } from '../util';
+
+const profile = async ( msg, user ) => {
+	const embed = new MessageEmbed();
+
+	embed.setColor( 'RANDOM' );
+	embed.setImage( user.displayAvatarURL() );
+	embed.setAuthor( msg.member.user.username, msg.member.user.displayAvatarURL() );
+	embed.setTitle( `${ user.username}#${ user.discriminator}` );
+	embed.setTimestamp( Date.now() );
+
+	if ( user.bot ) {
+		embed.setFooter( 'Bot' );
+	} else {
+		embed.setFooter( 'Human' );
+	}
+
+	return embed;
+};
 
 export default {
 	name: 'profile',
@@ -13,31 +31,14 @@ export default {
 	},
 	run: async ( _client, msg, args ) => {
 		let user;
-		let error = false;
 
 		if ( args[0] ) {
-			const id = args[0].replace( /[<]|!|@|[>]/g, '' );
-			user = await msg.guild.members.fetch( id ).catch( () => {
-				sendMsg( msg, 'El Usuario No Existe' );
-				error = true;
-			} );
-			if ( error ) return;
+			user = await getUserWithId( msg, args[0] );
 		}
+		if ( user === 'notFound' ) return sendMsg( msg, 'Usuario No Encontrado' );
 
 		const dataUser = user || msg.member;
-		const embed = new MessageEmbed();
-
-		embed.setColor( 'RANDOM' );
-		embed.setImage( dataUser.user.displayAvatarURL() );
-		embed.setAuthor( msg.member.user.username, msg.member.user.displayAvatarURL() );
-		embed.setTitle( `${dataUser.user.username}#${dataUser.user.discriminator}` );
-		embed.setTimestamp( Date.now() );
-
-		if ( dataUser.user.bot ) {
-			embed.setFooter( 'Bot' );
-		} else {
-			embed.setFooter( 'Human' );
-		}
+		const embed = await profile( msg, dataUser.user );
 
 		sendMsg( msg, embed );
 		msg.delete();
